Move starting video chat to menuAttachment
